+++ NEWS.md
@@ -60,7 +60,6 @@ for quite some time.
 + Adjusted cotton candy damage.
 + Fixed gustflowers spawning from inventory.
 + Made duckraptors scarier. 
-
 + Added Oolong Tea.
 + Added Spoiled Tea.
 + Added Chai Tea.
@@ -76,7 +75,9 @@ for quite some time.
 + Added Marshmallow Tea.
 + Added Ambrosia Tea.
 + Added Algae Tea.
-
++ Added Green Jelly crockpot recipe.
++ Added Red Jelly crockpot recipe.
++ Added Crystal Candy crockpot recipe.
 + Added beta foreshadowing.
 
 ### Update 2014-6-23 (alpha-0.0.7 release)